% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot the location of fatal injuries}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numerical represenatiion of the state}

\item{year}{The year value ot plot the fatal injuries}
}
\value{
A polygon map and plotted points
}
\description{
This is a function that creates a polygon map and plots the longitude and langitude coordinates of fatal injuries in a specified state and year in a map.
It reads a numeric value representing the state and a year value. It reads the csv file in the working directory
that holds the longitude and langitude of the incidents in the particular year. The funtion filters the incidents ny the given state
}
\details{
In the case of wrong state numerical representation, or zero fatal injuries for that state in the particular year, 
or there is no csv file in the current working directory under the specified year value, an error message is provided
}
\examples{
\dontrun{
fars_map_state(10, 2014)
fars_map_state(1, 2013)
}
}
