% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple csv files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector or list of year values}
}
\value{
A list of tibbles - one for each year
}
\description{
This is a function that reads multiple csv files in the standard format of \code{accident_YEAR.csv.bz2} into R.  
The function reads a vector or list of year values and calls \code{make_filename} to create the csv file names 
and then calls \code{fars_read} to read in the csv files
}
\details{
An error message is provided if no csv files are present in the current working directory under the 
year values that are provided,
}
\examples{
\dontrun{
fars_read_years(c("2014", "2015"))
fars_read_years(list("2014", "2015"))
}
}
