% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarise the fatal injuries}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector or list of year values}
}
\value{
A table of frequencies by year (columns) and month (rows)
}
\description{
This is a function that row-binds a list of tibbles and summarises the fatal injuries by Year and Month
}
\details{
An error message is provided if no csv files are present in the current working directory under the
year values that are provided,
}
\examples{
\dontrun{
fars_summarize_years(c("2014", "2015"))
fars_summarize_years(list("2014", "2015"))
}
}
