% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Make a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year as numeric or string}
}
\value{
A file name in string format
}
\description{
This is a function that concatenates string values and creates a filename in a standard format of \code{accident_YEAR.csv.bz2}
}
\details{
The function reads a value for the year and creates a filename in a standard format of \code{accident_YEAR.csv.bz2}
}
\examples{
\dontrun{
make_filename("2014")
make_filename(2014)
}
}
